chore(backend:auth)：JWT优化一些细节

+++ fitness-system/src/main/java/com/wsw/fitnesssystem/auth/infrastructure/token/model/RefreshTokenClaims.java
@@ -44,5 +44,5 @@ public class RefreshTokenClaims {
     private String username;
 
     /** Token 版本号，用于全局或单用户 Token 失效控制 */
-    private Integer tokenVersion;
+    private Long tokenVersion;
 }
